+++ backend/backend/urls.py
@@ -20,5 +20,6 @@ from django.urls import path, include
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('users/', include('users.urls')),
-    path('questions/', include('questions.urls'))
+    path('questions/', include('questions.urls')),
+    path('commitments', include('commitments.urls')),
 ]
