+++ backend/commitments/urls.py
@@ -1,0 +1,7 @@
+# users/urls.py
+from django.urls import path
+from . import views
+
+urlpatterns = [
+    path('<str:id>/', views.get_one, name='questions_view')
+]
