+++ backend/questions/models.py
@@ -12,13 +12,6 @@ class ModulesESG(DjangoCassandraModel):
     state = columns.Text(required=True)  # State of the module (e.g., 'open', 'validation', 'verified')
     calculated_score = columns.Double(required=False)  # Calculated score of the module in validation state
 
-class CommitmentPacts(DjangoCassandraModel):
-    id = columns.UUID(primary_key=True, default=uuid4)
-    id_client = columns.UUID(required=True)
-    creation_date = columns.DateTime(required=True)  # Date of the creation of the "Pacte d'engagement"
-    answers_commitments = columns.List(columns.UUID(required=True), required=True) # List of answers with isEngagement = True
-    calculated_score = columns.Double(required=True)  # Calculated score of the module in validation state
-
 class Answers(DjangoCassandraModel):
     id = columns.UUID(primary_key=True, default=uuid4)
     id_challenge = columns.UUID(required=True)
